# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StreamSphere Go API
  version: 1.0.0
  description: Contratos principales expuestos por el API Gateway.
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/videos/{id}/comments:
    get:
      summary: Listar comentarios
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        '200': { description: Comentarios paginados }
    post:
      summary: Publicar comentario
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string, maxLength: 1000 }
                parentId: { type: string }
      responses:
        '201': { description: Comentario publicado }
...